// client/src/components/commons/Button.ts
import styled from "styled-components";

export const Button = styled.button`
        background-color: var(--button-bc);
        color: white;
        border: none;
        padding: 9px 10px;
        border-radius: 3px;
        transition: all .2s;
        cursor: pointer;
        outline: none;
`;